refactor(bookmark): centralize bookmark toggle logic
- Extracted bookmark toggle logic into a new `useToggleBookmark` hook.
- Updated `poster-top-titles.tsx` to use the new hook.
- Updated `publication-detail-main/index.tsx` to use the new hook.
- Removed redundant bookmark-related imports and inline logic in components.
